Add missing viewport <meta> tag
viewport <meta> is required and without it, the media attribute won't work at all. Not only it is required, but it also has to be put before the media attribute is used because the browser’s preload scanner will not wait for other meta tags to be parsed before evaluating the media query in <link> tags.